Properly compute size of constant pool and include in emitted size.
Fixes #8.

Also added our first end-to-end (CLIF -> machine code) test to verify
that the constant pool is being emitted properly, with the correct
relative offset in the LDR instruction that references a constant. Fixed
the offset computation to actually be relative to the LDR, not the start
of the function's code.